requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
Pillow==10.0.1
stripe==7.6.0
python-dotenv==1.0.0
email-validator==2.0.0
bcrypt==4.0.1
psycopg2-binary==2.9.7 